requirements.txt
flask
python-dotenv
supabase
pymupdf
sentence-transformers
faiss-cpu
python-dateutil
google-generativeai